refactor: merge ARGuideViewModel extensions back into main file
Merged +CalibrationAndPractice (331 lines) and +ImmersiveAndTracking
(490 lines) back into ARGuideViewModel.swift (988 lines total).
Deleted both extension files.
